pyproject.toml
[tool.poetry]
name = "chat-server"
version = "0.1.0"
description = "A Chat Server"
readme = "README.md"
authors = ["Loki <yongxiang.zhan@outlook.com>"]
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.11"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
pydantic = "^2.5.2"
pymongo = "^4.11.3"
google-auth-oauthlib = "^1.2.1"
google-auth = "^2.38.0"
uvicorn = "^0.34.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
chat-server = "src.server:start"
